+++ behave/runner_util.py
@@ -6,11 +6,11 @@ Contains utility functions and classes for Runners.
 from behave import parser
 from behave.model import FileLocation
 from bisect import bisect
+from six import string_types
 import glob
 import os.path
 import re
 import sys
-import types
 
 
 # -----------------------------------------------------------------------------
@@ -278,7 +278,7 @@ def parse_features(feature_files, language=None):
     features = []
     for location in feature_files:
         if not isinstance(location, FileLocation):
-            assert isinstance(location, basestring)
+            assert isinstance(location, string_types)
             location = FileLocation(os.path.normpath(location))
 
         if location.filename == scenario_collector.filename:
@@ -351,7 +351,7 @@ def make_undefined_step_snippet(step, language=None):
     :param language: i18n language, optionally needed for step text parsing.
     :return: Undefined-step snippet (as string).
     """
-    if isinstance(step, types.StringTypes):
+    if isinstance(step, string_types):
         step_text = step
         steps = parser.parse_steps(step_text, language=language)
         step = steps[0]