+++ CHANGES.rst
@@ -20,6 +20,7 @@ NEWS and CHANGES:
     * ScenarioOutline: Annotates name with row.id, ... to better represent row.
     * NEW: Active Tags, see docs (`New and Noteworthy`_).
     * NEW: Test stages, see docs (`New and Noteworthy`_).
+    * NEW: User-specific configuration data, see docs (`New and Noteworthy`_).
 
   - Model:
 
@@ -31,10 +32,11 @@ NEWS and CHANGES:
 
 ENHANCEMENTS:
 
+  * issue #242: JUnitReporter can show scenario tags (provided by: rigomes)
   * issue #240: Test Stages with different step implementations (provided by: attilammagyar, jenisys)
   * issue #238: Allow to skip scenario in step function (provided by: hotgloupi, jenisys)
   * issue #228: Exclude scenario fron run (provided by: jdeppe, jenisys)
-  * issue #242: JUnitReporter can show scenario tags (provided by: rigomes)
+  * issue #227: Add a way to add command line options to behave (provided by: attilammagyar, jenisys)
 
 FIXED: